product-classification-enhanced/requirements.txt
# Core
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0

# Machine Learning
torch==2.0.1
torchvision==0.15.2
transformers==4.31.0
Pillow==10.0.0
numpy==1.24.3
scikit-learn==1.3.0

# Utilities
requests==2.31.0
pandas==2.0.3
joblib==1.3.1
tqdm==4.65.0

# Development (optional)
pytest==7.4.0
black==23.3.0
flake8==6.0.0
